refactor : 캐릭터 도감 조회 시 image url 전체 반환
