+++ src/main/java/univ/earthbreaker/namu/core/domain/common/Constant.java
@@ -1,6 +1,9 @@
-package univ.earthbreaker.namu.core.api;
+package univ.earthbreaker.namu.core.domain.common;
 
 public final class Constant {
+
+	private Constant() {
+	}
 
 	public static final String IMAGE_SYSTEM_PROPERTY = "image.access.url";
 	public static final String IMAGE_ACCESS_URL = System.getProperty(IMAGE_SYSTEM_PROPERTY);
